Check

+++ assets/shaders/standard.frag
@@ -30,7 +30,7 @@ struct Light {
 
 struct Camera {
     vec3 position;
-    vec2 resolution;
+    ivec2 resolution;
 };
 
 struct Environment {
